Moved exceptions to git.api
