+++ core/core-impl/src/test/java/io/fabric8/launcher/core/impl/MissionControlIT.java
@@ -16,7 +16,7 @@ import io.fabric8.launcher.core.api.ProjectileBuilder;
 import io.fabric8.launcher.service.github.api.GitHubRepository;
 import io.fabric8.launcher.service.github.api.GitHubService;
 import io.fabric8.launcher.service.github.api.GitHubServiceFactory;
-import io.fabric8.launcher.service.github.api.NoSuchRepositoryException;
+import io.fabric8.launcher.service.git.api.NoSuchRepositoryException;
 import io.fabric8.launcher.service.github.spi.GitHubServiceSpi;
 import io.fabric8.launcher.service.github.test.GitHubTestCredentials;
 import io.fabric8.launcher.service.openshift.api.OpenShiftProject;
