+++ services/git-service-api/src/main/java/io/fabric8/launcher/service/git/api/NoSuchRepositoryException.java
@@ -1,4 +1,4 @@
-package io.fabric8.launcher.service.github.api;
+package io.fabric8.launcher.service.git.api;
 
 /**
  * Indicates a specified repository does not exist.
